--- a/be/src/olap/rowset/segment_v2/inverted_index/query/conjunction_query.cpp
+++ b/be/src/olap/rowset/segment_v2/inverted_index/query/conjunction_query.cpp
@@ -37,15 +37,15 @@ ConjunctionQuery::~ConjunctionQuery() {
     }
 }
 
-void ConjunctionQuery::add(const std::wstring& field_name,
-                           const std::vector<std::wstring>& wterms) {
-    if (wterms.size() < 1) {
+void ConjunctionQuery::add(const std::wstring& field_name, const std::vector<std::string>& terms) {
+    if (terms.size() < 1) {
         _CLTHROWA(CL_ERR_IllegalArgument, "ConjunctionQuery::add: terms.size() < 1");
     }
 
     std::vector<TermIterator> iterators;
-    for (auto& wterm : wterms) {
-        Term* t = _CLNEW Term(field_name.c_str(), wterm.c_str());
+    for (auto& term : terms) {
+        std::wstring ws_term = StringUtil::string_to_wstring(term);
+        Term* t = _CLNEW Term(field_name.c_str(), ws_term.c_str());
         _terms.push_back(t);
         TermDocs* term_doc = _reader->termDocs(t);
         _term_docs.push_back(term_doc);