--- a/be/src/olap/rowset/segment_v2/inverted_index/query/conjunction_query.h
+++ b/be/src/olap/rowset/segment_v2/inverted_index/query/conjunction_query.h
@@ -38,7 +38,7 @@ public:
         _conjunction_ratio = conjunction_ratio;
     }
 
-    void add(const std::wstring& field_name, const std::vector<std::wstring>& wterms);
+    void add(const std::wstring& field_name, const std::vector<std::string>& terms);
     void search(roaring::Roaring& roaring);
 
 private: