--- a/be/src/vec/functions/match.cpp
+++ b/be/src/vec/functions/match.cpp
@@ -102,26 +102,32 @@ inline doris::segment_v2::InvertedIndexQueryType FunctionMatchBase::get_query_ty
     return doris::segment_v2::InvertedIndexQueryType::UNKNOWN_QUERY;
 }
 
-inline std::vector<std::wstring> FunctionMatchBase::analyse_data_token(
+inline std::vector<std::string> FunctionMatchBase::analyse_data_token(
         const std::string& column_name, InvertedIndexCtx* inverted_index_ctx,
         const ColumnString* string_col, int32_t current_block_row_idx,
         const ColumnArray::Offsets64* array_offsets, int32_t& current_src_array_offset) {
-    std::vector<std::wstring> data_tokens;
+    std::vector<std::string> data_tokens;
     auto query_type = get_query_type_from_fn_name();
     if (array_offsets) {
         for (auto next_src_array_offset = (*array_offsets)[current_block_row_idx];
              current_src_array_offset < next_src_array_offset; ++current_src_array_offset) {
             const auto& str_ref = string_col->get_data_at(current_src_array_offset);
-            std::vector<std::wstring> element_tokens =
+            auto reader = doris::segment_v2::InvertedIndexReader::create_reader(
+                    inverted_index_ctx, str_ref.to_string());
+
+            std::vector<std::string> element_tokens =
                     doris::segment_v2::InvertedIndexReader::get_analyse_result(
-                            column_name, str_ref.to_string(), query_type, inverted_index_ctx,
+                            reader.get(), inverted_index_ctx->analyzer, column_name, query_type,
                             false);
             data_tokens.insert(data_tokens.end(), element_tokens.begin(), element_tokens.end());
         }
     } else {
         const auto& str_ref = string_col->get_data_at(current_block_row_idx);
+        auto reader = doris::segment_v2::InvertedIndexReader::create_reader(inverted_index_ctx,
+                                                                            str_ref.to_string());
+
         data_tokens = doris::segment_v2::InvertedIndexReader::get_analyse_result(
-                column_name, str_ref.to_string(), query_type, inverted_index_ctx, false);
+                reader.get(), inverted_index_ctx->analyzer, column_name, query_type, false);
     }
     return data_tokens;
 }
@@ -138,10 +144,12 @@ Status FunctionMatchAny::execute_match(const std::string& column_name,
     }
     VLOG_DEBUG << "begin to run FunctionMatchAny::execute_match, parser_type: "
                << inverted_index_parser_type_to_string(parser_type);
-    std::vector<std::wstring> query_tokens =
+    auto reader = doris::segment_v2::InvertedIndexReader::create_reader(inverted_index_ctx,
+                                                                        match_query_str);
+    std::vector<std::string> query_tokens =
             doris::segment_v2::InvertedIndexReader::get_analyse_result(
-                    column_name, match_query_str,
-                    doris::segment_v2::InvertedIndexQueryType::MATCH_ANY_QUERY, inverted_index_ctx);
+                    reader.get(), inverted_index_ctx->analyzer, column_name,
+                    doris::segment_v2::InvertedIndexQueryType::MATCH_ANY_QUERY);
     if (query_tokens.empty()) {
         LOG(WARNING) << fmt::format(
                 "token parser result is empty for query, "
@@ -152,7 +160,7 @@ Status FunctionMatchAny::execute_match(const std::string& column_name,
 
     auto current_src_array_offset = 0;
     for (int i = 0; i < input_rows_count; i++) {
-        std::vector<std::wstring> data_tokens =
+        std::vector<std::string> data_tokens =
                 analyse_data_token(column_name, inverted_index_ctx, string_col, i, array_offsets,
                                    current_src_array_offset);
 
@@ -181,10 +189,12 @@ Status FunctionMatchAll::execute_match(const std::string& column_name,
     }
     VLOG_DEBUG << "begin to run FunctionMatchAll::execute_match, parser_type: "
                << inverted_index_parser_type_to_string(parser_type);
-    std::vector<std::wstring> query_tokens =
+    auto reader = doris::segment_v2::InvertedIndexReader::create_reader(inverted_index_ctx,
+                                                                        match_query_str);
+    std::vector<std::string> query_tokens =
             doris::segment_v2::InvertedIndexReader::get_analyse_result(
-                    column_name, match_query_str,
-                    doris::segment_v2::InvertedIndexQueryType::MATCH_ALL_QUERY, inverted_index_ctx);
+                    reader.get(), inverted_index_ctx->analyzer, column_name,
+                    doris::segment_v2::InvertedIndexQueryType::MATCH_ALL_QUERY);
     if (query_tokens.empty()) {
         LOG(WARNING) << fmt::format(
                 "token parser result is empty for query, "
@@ -195,7 +205,7 @@ Status FunctionMatchAll::execute_match(const std::string& column_name,
 
     auto current_src_array_offset = 0;
     for (int i = 0; i < input_rows_count; i++) {
-        std::vector<std::wstring> data_tokens =
+        std::vector<std::string> data_tokens =
                 analyse_data_token(column_name, inverted_index_ctx, string_col, i, array_offsets,
                                    current_src_array_offset);
 
@@ -230,11 +240,12 @@ Status FunctionMatchPhrase::execute_match(const std::string& column_name,
     }
     VLOG_DEBUG << "begin to run FunctionMatchPhrase::execute_match, parser_type: "
                << inverted_index_parser_type_to_string(parser_type);
-    std::vector<std::wstring> query_tokens =
+    auto reader = doris::segment_v2::InvertedIndexReader::create_reader(inverted_index_ctx,
+                                                                        match_query_str);
+    std::vector<std::string> query_tokens =
             doris::segment_v2::InvertedIndexReader::get_analyse_result(
-                    column_name, match_query_str,
-                    doris::segment_v2::InvertedIndexQueryType::MATCH_PHRASE_QUERY,
-                    inverted_index_ctx);
+                    reader.get(), inverted_index_ctx->analyzer, column_name,
+                    doris::segment_v2::InvertedIndexQueryType::MATCH_PHRASE_QUERY);
     if (query_tokens.empty()) {
         LOG(WARNING) << fmt::format(
                 "token parser result is empty for query, "
@@ -245,7 +256,7 @@ Status FunctionMatchPhrase::execute_match(const std::string& column_name,
 
     auto current_src_array_offset = 0;
     for (int i = 0; i < input_rows_count; i++) {
-        std::vector<std::wstring> data_tokens =
+        std::vector<std::string> data_tokens =
                 analyse_data_token(column_name, inverted_index_ctx, string_col, i, array_offsets,
                                    current_src_array_offset);