## Proposed changes

Issue Number: close #xxx

improve performance for match without index, from 2 min 34.54 sec to 9.16s, on httplogs dataset.
```
mysql> select COUNT() from wc_httplogs_without_index where request match "GET";                                                                  +-----------+
| count(*)  |
+-----------+
| 245552656 |
+-----------+
1 row in set (29.29 sec)

mysql> select COUNT() from wc_httplogs where request match "GET";
+-----------+
| count(*)  |
+-----------+
| 245552656 |
+-----------+
1 row in set (0.51 sec)

mysql> select COUNT() from wc_httplogs_without_index where request match "GET";
+-----------+
| count(*)  |
+-----------+
| 245552656 |
+-----------+
1 row in set (2 min 34.54 sec)

mysql> select COUNT() from wc_httplogs_without_index where request match "GET";                                                                 +-----------+
| count(*)  |
+-----------+
| 245552656 |
+-----------+
1 row in set (9.16 sec)

```
## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...