@airborne12 2023-09-21
run buildall