@airborne12 2023-09-22
run buildall